# docs/api.yml
openapi: 3.0.3
info:
  title: TD2 Wallet
  version: 1.0.0
  description: STD22098 - STD22106
servers:
  - url: http://localhost:8080
...
paths:
...
  /transactions:
    get:
      tags:
        - Transaction
      summary: get all Transaction 
      description: Return a list of transaction 
      operationId: getTransactions
      parameters:
        - in: query
          name: queryPagination
          $ref: '#/components/parameters/queryPagination'
      responses:
        '200':
          $ref: '#/components/responses/TransactionListSuccess'
  /transaction/{id}:
    get:
      tags:
        - Transaction
      summary: get a transaction by id
      operationId: getTransactionById
      parameters:
        - in: path
          name: id 
          description: get transaction by id 
          required: true
          schema: 
            type: string
      responses:
        '200': 
          $ref: '#/components/responses/TransactionSuccess'
  /transaction:
    put:
      tags:
        - Transaction
      summary: update or create a transaction 
      operationId: updateTransaction
      requestBody:
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/Transaction'
      responses:
        '200': 
          $ref: '#/components/responses/TransactionSuccess'
  /transaction/{uid}:
    delete:
      tags:
        - Transaction
      summary: delete a transaction by id
      operationId: deleteTransactionById
      parameters:
        - in: path
          name: uid 
          description: delete transaction by id 
          required: true
          schema: 
            type: string
      responses:
        '200': 
          $ref: '#/components/responses/TransactionSuccess'
...
components:
  schemas:
    User:
      type: object
      properties:
        id: 
          type: string
        username:
          type: string
        email: 
          type: string
        password: 
          type: string
    Account:
      type: object
      properties:
        id: 
          type: integer
        type:
          type: string
          enum:
            - Banque
            - Espece
            - Mobile Money
        name: 
          type: string
          example: current account
        solde:
          type: integer
        user: 
          $ref: '#/components/schemas/User'
        currency: 
          $ref: '#/components/schemas/Currency'
        transaction: 
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    Currency:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: Euro
        code:
          type: string
          example: EUR
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum:
            - Debit
            - Credit
        label: 
          type: string
        amount:
          type: integer
        transaction_date:
          type: string
          format: date
        account: 
          $ref: '#/components/schemas/Account'
    Exception:
      type: object
      properties:
        status:
          type: object
          properties:
            status_code:
              type: string
              example: 200
            message:
              type: string
              example: OK
        body: 
          type: object
...
    TransactionListBody:
      allOf:
        - $ref: '#/components/schemas/Exception'
        - properties:
            body:
              type: array
              items: 
                $ref: '#/components/schemas/Account'
    TransactionResponse:
      allOf:
        - $ref: '#/components/schemas/Exception'
        - properties:
            body:
              $ref: '#/components/schemas/Transaction'
...
  parameters:
    queryPagination:
      name: queryPagination
      in: query
      description: for query pagination
      required: false
      schema:
        type: object
        properties:
          page:
            type: integer
            default: 1
          size:
            type: integer
            default: 50
...
  responses:
...
    TransactionListSuccess:
      description: list of transaction
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/TransactionListBody'
    TransactionSuccess: 
      description: return a transaction 
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/TransactionResponse'
...